Mmove migration code to _migration.go file and exxtract tests to _migartion_test.go file
