+++ github/resource_github_actions_secret.go
@@ -292,15 +292,3 @@ func encryptPlaintext(plaintext, publicKeyB64 string) ([]byte, error) {
 
 	return cipherText, nil
 }
-
-func resourceGithubActionsSecretInstanceStateUpgradeV0(ctx context.Context, rawState map[string]any, meta any) (map[string]any, error) {
-	log.Printf("[DEBUG] GitHub Actions Secret State before migration: %#v", rawState)
-	// Add the destroy_on_drift field with default value true if it doesn't exist
-	if _, ok := rawState["destroy_on_drift"]; !ok {
-		rawState["destroy_on_drift"] = true
-	}
-
-	log.Printf("[DEBUG] GitHub Actions Secret State after migration: %#v", rawState)
-
-	return rawState, nil
-}
